+++ Assets/Scripts/Weapons/Gun.cs
@@ -18,7 +18,7 @@ public class Gun : Weapon
         {
             if (Input.GetMouseButtonDown(0) && fireTime >= fireRate && !isReloading)
             {
-            Shoot((int)(damageType));
+            Shoot(damageType);
             }
             Reload();
             if (fireTime < fireRate)
